+++ lib/overlay_draggable_button.dart
@@ -13,7 +13,9 @@ showDebugBtn(BuildContext context, {Widget? button, Color? btnColor}) {
   ///widget第一次渲染完成
   WidgetsBinding.instance!.addPostFrameCallback((_) {
     dismissDebugBtn();
-    itemEntry = OverlayEntry(builder: (BuildContext context) => button ?? DraggableButtonWidget(btnColor: btnColor));
+    itemEntry = OverlayEntry(
+        builder: (BuildContext context) =>
+            button ?? DraggableButtonWidget(btnColor: btnColor));
 
     ///显示悬浮menu
     Overlay.of(context)?.insert(itemEntry!);