+++ lib/utils/copy_clipboard.dart
@@ -3,7 +3,8 @@ import 'package:flutter/services.dart';
 
 ///复制到粘贴板
 copyClipboard(BuildContext context, String? value) {
-  var snackBar = SnackBar(content: Text('copy $value to clipboard'));
+  var snackBar =
+      SnackBar(content: Text('$value\n\n copy success to clipboard'));
   ScaffoldMessenger.of(context).showSnackBar(snackBar);
-  Clipboard.setData(ClipboardData(text: value));
+  Clipboard.setData(ClipboardData(text: value ?? 'null'));
 }